+++ src/api/graphics/vulkan/lvn_vulkan.h
@@ -53,10 +53,10 @@ namespace lvn
 	void vksImplRenderCmdBeginFrameBuffer(LvnWindow* window, LvnFrameBuffer* frameBuffer);
 	void vksImplRenderCmdEndFrameBuffer(LvnWindow* window, LvnFrameBuffer* frameBuffer);
 
-	void vksImplBufferUpdateVertexData(LvnBuffer* buffer, void* vertices, uint32_t size, uint32_t offset);
-	void vksImplBufferUpdateIndexData(LvnBuffer* buffer, uint32_t* indices, uint32_t size, uint32_t offset);
-	void vksImplBufferResizeVertexBuffer(LvnBuffer* buffer, uint32_t size);
-	void vksImplBufferResizeIndexBuffer(LvnBuffer* buffer, uint32_t size);
+	void vksImplBufferUpdateVertexData(LvnBuffer* buffer, void* vertices, uint64_t size, uint64_t offset);
+	void vksImplBufferUpdateIndexData(LvnBuffer* buffer, uint32_t* indices, uint64_t size, uint64_t offset);
+	void vksImplBufferResizeVertexBuffer(LvnBuffer* buffer, uint64_t size);
+	void vksImplBufferResizeIndexBuffer(LvnBuffer* buffer, uint64_t size);
 	void vksImplUpdateUniformBufferData(LvnWindow* window, LvnUniformBuffer* uniformBuffer, void* data, uint64_t size);
 	void vksImplUpdateDescriptorSetData(LvnDescriptorSet* descriptorSet, LvnDescriptorUpdateInfo* pUpdateInfo, uint32_t count);
 	LvnTexture* vksImplFrameBufferGetImage(LvnFrameBuffer* frameBuffer, uint32_t attachmentIndex);
